Small improvements in the odata tests
